Fix configureTracking session replay startup
